Update unit tests to look for legacy topic references in `References`
Legacy topic references are attributes whose `key` end with `Id` and `Value` refers to a `UniqueKey`. They were stored as attributes because there wasn't another way to establish a keyed reference to a single topic. That is no longer the case now that we have full support for `Topic.References`. As such, on `Import()`, these are now migrated to `TopicData.References` and imported into `Topic.References` (6d060d0).

As part of this, legacy topic references that don't pass initial validation are now imported as attributes—where as, previously, they were skipped. This helps avoid accidental data loss, and especially for situations where the topic reference may have been a false positive. (A risk with using this option previously was that it might mistake attributes that just happen to fit the pattern, but were never intended as topic references.)

To support this, the unit tests need to be updated to a) look for the legacy topic references in `Topic.References`, and b) expect that unmapped or orphaned topic references will continue to show up in `Topic.Attributes`, instead of being deleted.