Rename `AttributeData` to `RecordData`
This better maps to the new `TrackedRecord<T>` class in OnTopic, which is the foundation for both `Topic.Attributes` and `Topic.References`, which corresponds to the use in the OnTopic Data Transfer Library, where `RecordData` is now the foundation for both `TopicData.Attributes` and `TopicData.References`.

This includes updating unit test and variables associated with `RecordData` instances.