Avoid setters on collection-valued properties
Properties with a return type of a collection should not have a setter. That's because collections should be updated directly, using the collection semantics, instead of replaced entirely via the parent property. Unfortunately, we can't remove the setter entirely because it is relied upon for .NET deserialization. We can, however, help mitigate the risk by restricting the setting of the collection to `init`.

This still trips the `CA2227` warning in code analysis. I'd consider that a bug. But, regardless, we'll need to maintain the suppression in the `csproj` file.