Conditionally expose setter for deprecated `DerivedTopicKey`
The `DerivedTopicKey` property is deprecated. Unfortunately, however, we still require it in order to deserialize JSON exported by legacy versions of the OnTopic Data Transfer library. Ideally, we'd hide the entire implementation from the public interface, and only support it for deserialization. Since that's not practical, we can instead at least prevent it from being set externally, or serialized if null.