Test conditional writing of `DerivedTopicKey` in .NET 5
With .NET 5, we added a `[JsonIgnore()]` condition to the `DerivedTopicKey` so that it's not serialized to JSON if the value is `null` (960f1bd)—which it will always be after an `Export()` since the extension method no longer writes the `BaseTopic.Key` to `DerivedTopicKey`. This condition within the unit test evaluates that to ensure it works properly.