Introduced unit test for verifying backward compatibility
The `KeyValuesPairConverter` is intended to assure backward compatibility with the legacy `RelationshipData` object which wrote associated unique keys into a `Relationships` array instead of a `Values` array. On deserialization, JSON with the `Relationships` array should be successfully read into the `Values` collection. This unit test verifies that is the case.