Hard-coded date time values to avoid false negatives in unit tests
The JSON serializer/deserializer will truncate trailing `0`s, whereas .NET's string conversion will not. This leads to false negatives in the unit tests. This doesn't affect actual serialization or deserialization, since the ticks are fully represented once converted back to a `DateTime` instance, but it interferes with the unit tests where the `expected` JSON is being manually constructed for exact comparison.

To mitigate this, a `lastModified` date is hard-coded into the unit tests, and then formatted as a short string to ensure that it doesn't include the milliseconds. This avoids scenarios where there are discrepancies in the data.