Remove multi-targeting of .NET 5
In a previous release, we updated the OnTopic Data Transfer library to multi-target .NET 5 (2f8ca6b). That was premature. The primary driver of this was the ability to use new .NET 5 attributes for (de)serialization. But those can be accessed by simply upgrading the `System.Text.Json` package to 5.0.0 (439b41c); that doesn't actually necessitate the use of .NET 5 itself. Given that, we can undo the multi-targeting previously introduced.