Configure solution to auto-generate XML documentation
We are diligent about maintaining XML Docs for all types and members. But we don't currently _do_ anything with that documentation. By auto-generating it on compilation, we at least ensure that it will end up being included in NuGet packages—and thus available to IntelliSense for implementers. In addition, this will allow us to use e.g. DocFX in the future as part of our CI/CD process, should we choose.

As part of this, accept the default XML output path. This is the same as the one we have configured, and there's no real benefit to being explicit in this case.